move samples to V2
