+++ sdk/metricsadvisor/ai-metrics-advisor/package.json
@@ -50,14 +50,14 @@
     "build:autorest": "autorest ./swagger/README.md --typescript --version=3.0.6267",
     "build:browser": "tsc -p . && cross-env ONLY_BROWSER=true rollup -c 2>&1",
     "build:node": "tsc -p . && cross-env ONLY_NODE=true rollup -c 2>&1",
-    "build:samples": "dev-tool samples prep && cd dist-samples && tsc -p .",
+    "build:samples": "echo Obsolete.",
     "build:test": "tsc -p . && rollup -c 2>&1",
     "build": "tsc -p . && rollup -c 2>&1 && api-extractor run --local",
-    "check-format": "prettier --list-different --config ../../../.prettierrc.json --ignore-path ../../../.prettierignore \"src/**/*.ts\" \"test/**/*.ts\" \"*.{js,json}\"",
+    "check-format": "prettier --list-different --config ../../../.prettierrc.json --ignore-path ../../../.prettierignore \"src/**/*.ts\" \"test/**/*.ts\" \"samples-dev/**/*.ts\" \"*.{js,json}\"",
     "clean": "rimraf dist dist-esm dist-browser dist-test dist-test temp types *.tgz *.log",
-    "execute:samples": "npm run build:samples && dev-tool samples run samples/javascript/ && dev-tool samples run dist-samples/typescript/dist/dist-samples/typescript/src/",
+    "execute:samples": "dev-tool samples run samples-dev",
     "extract-api": "tsc -p . && api-extractor run --local",
-    "format": "prettier --write --config ../../../.prettierrc.json --ignore-path ../../../.prettierignore \"src/**/*.ts\" \"test/**/*.ts\" \"*.{js,json}\"",
+    "format": "prettier --write --config ../../../.prettierrc.json --ignore-path ../../../.prettierignore \"src/**/*.ts\" \"test/**/*.ts\" \"samples-dev/**/*.ts\" \"*.{js,json}\"",
     "integration-test:browser": "karma start --single-run",
     "integration-test:node": "nyc mocha -r esm --require source-map-support/register --reporter ../../../common/tools/mocha-multi-reporter.js --full-trace -t 300000  \"dist-esm/test/internal/*.spec.js\" \"dist-esm/test/public/*.spec.js\" \"dist-esm/test/public/node/*.spec.js\" \"dist-esm/test/internal/node/*.spec.js\"",
     "integration-test": "npm run integration-test:node && npm run integration-test:browser",
@@ -126,7 +126,7 @@
     "typedoc": "0.15.2"
   },
   "//sampleConfiguration": {
-    "skip": [
+     "skip": [
       "alertingConfig.js",
       "dataFeed.js",
       "detectionConfig.js",
@@ -137,6 +137,18 @@
       "metricQueries.js",
       "quickstart.js",
       "seriesData.js"
-    ]
+    ],
+     "productName": "Azure Metrics Advisor",
+      "productSlugs": [
+      "azure",
+      "azure-cognitive-services",
+      "azure-metrics-advisor"
+    ],
+      "requiredResources": {
+      "Azure Cognitive Services account": "https://docs.microsoft.com/azure/cognitive-services/cognitive-services-apis-create-account"
+    },
+    "skipFolder": false,
+    "disableDocsMs": false,
+    "apiRefLink": "https://docs.microsoft.com/javascript/api/"
   }
 }
