+++ sdk/metricsadvisor/ai-metrics-advisor/samples-dev/hooks.ts
@@ -1,3 +1,11 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+/**
+ * @summary This sample demonstrates Metrics Advisor Hooks CRUD operations.
+ */
+// Load the .env file if it exists
+
 import * as dotenv from "dotenv";
 dotenv.config();
 import {
