+++ sdk/metricsadvisor/ai-metrics-advisor/samples-dev/quickstart.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 /**
- * This sample demonstrates how to get started by creating a data feed, checking ingestion status,
+ *  @summary This sample demonstrates how to get started by creating a data feed, checking ingestion status,
  * creating detection and alerting configurations, and querying for alerts and anomalies.
  */
 // Load the .env file if it exists
