--- sdk/metricsadvisor/ai-metrics-advisor/samples/javascript/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "azure-ai-metrics-advisor-samples-js",
-  "private": true,
-  "version": "0.1.0",
-  "description": "Azure Cognitive Services Metrics Advisor client library samples for JavaScript",
-  "engine": {
-    "node": ">=8.0.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Azure/azure-sdk-for-js.git"
-  },
-  "keywords": [
-    "Azure",
-    "Cognitive Services",
-    "Metrics Advisor",
-    "Node.js",
-    "JavaScript"
-  ],
-  "author": "Microsoft Corporation",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Azure/azure-sdk-for-js/issues"
-  },
-  "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/master/sdk/metricsadvisor/ai-metrics-advisor",
-  "sideEffects": false,
-  "dependencies": {
-    "@azure/ai-metrics-advisor": "latest",
-    "@azure/identity": "latest",
-    "dotenv": "^8.2.0"
-  }
-}
