--- sdk/metricsadvisor/ai-metrics-advisor/samples/javascript/sample.env
@@ -1,15 +1,0 @@
-METRICS_ADVISOR_ENDPOINT=https://<resource name>.cognitiveservices.azure.com/
-METRICS_ADVISOR_SUBSCRIPTION_KEY=<subscription key>
-METRICS_ADVISOR_API_KEY=<api key>
-
-METRICS_ADVISOR_DETECTION_CONFIG_ID=<id of an existing detection configuration>
-METRICS_ADVISOR_ALERT_CONFIG_ID=<id of an existing alert configuration>
-
-METRICS_ADVISOR_ALERT_ID=<id of an alert>
-METRICS_ADVISOR_INCIDENT_ID=<id of an incident>
-METRICS_ADVISOR_DATAFEED_ID=<id of an existing data feed>
-METRICS_ADVISOR_METRIC_ID=<id of an existing metric>
-METRICS_ADVISOR_HOOK_ID=<id of an existing hook>
-
-METRICS_ADVISOR_AZURE_SQL_SERVER_CONNECTION_STRING=<connection string to SQL Server>
-METRICS_ADVISOR_AZURE_SQL_SERVER_QUERY=<SQL query to SQL Server data>
