--- sdk/metricsadvisor/ai-metrics-advisor/samples/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "outDir": "typescript/dist",
-    "lib": ["ES6", "dom"]
-  },
-  "include": ["typescript/src/**.ts"],
-  "exclude": ["typescript/*.json", "**/node_modules/", "../node_modules", "../types"]
-}
