--- sdk/metricsadvisor/ai-metrics-advisor/samples/typescript/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "moduleResolution": "node",
-
-    "allowSyntheticDefaultImports": true,
-
-    "strict": true,
-    "alwaysStrict": true,
-
-    "outDir": "dist",
-    "rootDir": "src"
-  },
-  "include": ["src/**.ts"],
-  "exclude": ["node_modules"]
-}
