+++ sdk/metricsadvisor/ai-metrics-advisor/tsconfig.json
@@ -2,7 +2,12 @@
   "extends": "../../../tsconfig.package",
   "compilerOptions": {
     "outDir": "./dist-esm",
-    "declarationDir": "./types"
+    "declarationDir": "./types",
+    "paths": {
+      "@azure/ai-metrics-advisor": [
+        "./src/index"
+      ]
+    }
   },
   "exclude": [
     "node_modules",
@@ -14,5 +19,10 @@
     "./samples/**/*.ts",
     "./dist-samples/**/*.ts",
     "./dom-shims.d.ts"
+  ],
+  "include": [
+    "src/**/*.ts",
+    "test/**/*.ts",
+    "samples-dev/**/*.ts"
   ]
 }
